[STEP 3] test: Repository 관련 테스트 코드 작성
